Add story and person tags
